[GLT-4230] Make Target Cell Recovery an Integer

+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/SampleStockSingleCell.java
@@ -6,9 +6,9 @@ public interface SampleStockSingleCell extends SampleStock {
 
   public static final String SUBCATEGORY_NAME = "Single Cell (stock)";
 
-  public BigDecimal getTargetCellRecovery();
+  public Integer getTargetCellRecovery();
 
-  public void setTargetCellRecovery(BigDecimal targetCellRecovery);
+  public void setTargetCellRecovery(Integer targetCellRecovery);
 
   public BigDecimal getCellViability();
